LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/instantiation-error-1-expected.txt

CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.

FAIL Test that missing exports lead to SyntaxError events on window and load events on script, and that exceptions are remembered null is not an object (evaluating 'exn.constructor')
